+++ cognite/neat/graph/transformers/_classic_cdf.py
@@ -1,6 +1,6 @@
 from typing import cast
 
-from rdflib import Graph, Literal, URIRef
+from rdflib import RDF, Graph, Literal, URIRef
 
 from cognite.neat.constants import DEFAULT_NAMESPACE
 from cognite.neat.graph import extractors
@@ -21,11 +21,16 @@ class AddAssetDepth(BaseTransformer):
                              <{asset_id}> <{root_prop}> ?root .}}"""
 
     def __init__(
-        self, asset_type: URIRef | None = None, root_prop: URIRef | None = None, parent_prop: URIRef | None = None
+        self,
+        asset_type: URIRef | None = None,
+        root_prop: URIRef | None = None,
+        parent_prop: URIRef | None = None,
+        depth_typing: dict[int, str] | None = None,
     ):
         self.asset_type = asset_type or DEFAULT_NAMESPACE.Asset
         self.root_prop = root_prop or DEFAULT_NAMESPACE.root
         self.parent_prop = parent_prop or DEFAULT_NAMESPACE.parent
+        self.depth_typing = depth_typing
 
     def transform(self, graph: Graph) -> None:
         """Adds depth of asset in the asset hierarchy to the graph."""
@@ -33,6 +38,13 @@ class AddAssetDepth(BaseTransformer):
             asset_id = cast(tuple, result)[0]
             if depth := self.get_depth(graph, asset_id, self.root_prop, self.parent_prop):
                 graph.add((asset_id, DEFAULT_NAMESPACE.depth, Literal(depth)))
+
+                if self.depth_typing and (type_ := self.depth_typing.get(depth, None)):
+                    # remove existing type
+                    graph.remove((asset_id, RDF.type, None))
+
+                    # add new type
+                    graph.add((asset_id, RDF.type, DEFAULT_NAMESPACE[type_]))
 
     @classmethod
     def get_depth(
@@ -62,7 +74,12 @@ class AddAssetDepth(BaseTransformer):
 class AssetTimeSeriesConnector(BaseTransformer):
     description: str = "Connects assets to timeseries, thus forming bi-directional connection"
     _use_only_once: bool = True
-    _need_changes = frozenset({str(extractors.AssetsExtractor.__name__), str(extractors.TimeSeriesExtractor.__name__)})
+    _need_changes = frozenset(
+        {
+            str(extractors.AssetsExtractor.__name__),
+            str(extractors.TimeSeriesExtractor.__name__),
+        }
+    )
     _asset_template: str = """SELECT ?asset_id WHERE {{
                               <{timeseries_id}> <{asset_prop}> ?asset_id .
                               ?asset_id a <{asset_type}>}}"""
@@ -100,7 +117,12 @@ class AssetTimeSeriesConnector(BaseTransformer):
 class AssetSequenceConnector(BaseTransformer):
     description: str = "Connects assets to sequences, thus forming bi-directional connection"
     _use_only_once: bool = True
-    _need_changes = frozenset({str(extractors.AssetsExtractor.__name__), str(extractors.SequencesExtractor.__name__)})
+    _need_changes = frozenset(
+        {
+            str(extractors.AssetsExtractor.__name__),
+            str(extractors.SequencesExtractor.__name__),
+        }
+    )
     _asset_template: str = """SELECT ?asset_id WHERE {{
                               <{sequence_id}> <{asset_prop}> ?asset_id .
                               ?asset_id a <{asset_type}>}}"""
@@ -138,7 +160,12 @@ class AssetSequenceConnector(BaseTransformer):
 class AssetFileConnector(BaseTransformer):
     description: str = "Connects assets to files, thus forming bi-directional connection"
     _use_only_once: bool = True
-    _need_changes = frozenset({str(extractors.AssetsExtractor.__name__), str(extractors.FilesExtractor.__name__)})
+    _need_changes = frozenset(
+        {
+            str(extractors.AssetsExtractor.__name__),
+            str(extractors.FilesExtractor.__name__),
+        }
+    )
     _asset_template: str = """SELECT ?asset_id WHERE {{
                               <{file_id}> <{asset_prop}> ?asset_id .
                               ?asset_id a <{asset_type}>}}"""
@@ -174,7 +201,12 @@ class AssetFileConnector(BaseTransformer):
 class AssetEventConnector(BaseTransformer):
     description: str = "Connects assets to events, thus forming bi-directional connection"
     _use_only_once: bool = True
-    _need_changes = frozenset({str(extractors.AssetsExtractor.__name__), str(extractors.EventsExtractor.__name__)})
+    _need_changes = frozenset(
+        {
+            str(extractors.AssetsExtractor.__name__),
+            str(extractors.EventsExtractor.__name__),
+        }
+    )
     _asset_template: str = """SELECT ?asset_id WHERE {{
                               <{event_id}> <{asset_prop}> ?asset_id .
                               ?asset_id a <{asset_type}>}}"""
@@ -211,7 +243,10 @@ class AssetRelationshipConnector(BaseTransformer):
     description: str = "Connects assets via relationships"
     _use_only_once: bool = True
     _need_changes = frozenset(
-        {str(extractors.AssetsExtractor.__name__), str(extractors.RelationshipsExtractor.__name__)}
+        {
+            str(extractors.AssetsExtractor.__name__),
+            str(extractors.RelationshipsExtractor.__name__),
+        }
     )
     _asset_template: str = """SELECT ?source ?target WHERE {{
                               <{relationship_id}> <{relationship_source_xid_prop}> ?source_xid .
@@ -256,8 +291,20 @@ class AssetRelationshipConnector(BaseTransformer):
                 # files can be connected to multiple assets in the graph
                 for source_asset_id, target_asset_id in cast(list[tuple], assets_id_res):
                     # create a relationship between the two assets
-                    graph.add((source_asset_id, DEFAULT_NAMESPACE.relationship, relationship_id))
-                    graph.add((target_asset_id, DEFAULT_NAMESPACE.relationship, relationship_id))
+                    graph.add(
+                        (
+                            source_asset_id,
+                            DEFAULT_NAMESPACE.relationship,
+                            relationship_id,
+                        )
+                    )
+                    graph.add(
+                        (
+                            target_asset_id,
+                            DEFAULT_NAMESPACE.relationship,
+                            relationship_id,
+                        )
+                    )
 
                     # add source and target to the relationship
                     graph.add((relationship_id, DEFAULT_NAMESPACE.source, source_asset_id))